Moved WebUI class to ui submodule; renamed WebUI class
